Refactored some code from checkout
